[DEVX-1176] RateMatrix -> FxPlugin::RateMatrix
