+++ lib/global_reach_partners/fx_plugin.rb
@@ -1,2 +1,3 @@
 require 'global_reach_partners/fx_plugin/request'
+require 'global_reach_partners/fx_plugin/rate_matrix'
 require 'global_reach_partners/fx_plugin/deal'
