+++ spec/fixtures/cassettes/fx_plugin/rate_matrix/fetch.yml
