+++ spec/fixtures/cassettes/fx_plugin/rate_matrix/get_rates.yml
